Episode Page Update
Added the Episode's Date to the Episode Page as well. To isolate what code is running on what page (since the Extension's pattern matching matches all subpages of any given Show), the extension was broken into classes which are initialized when appropriate for the given page. Those classes are separated into different files for clarity.