Fix Funi Changes
Changes to Funi's page data broke extension for the Show Homepage.
Episodes no longer seem to have their own page ( video loads in pseudo-full screen with no apparent location for the Episode Information). No changes made to dater_episode as it is not throwing an error and there is no way to determine if it needs to/how it can be fixed.